feat(clients): replace History Book notes with this client's activity feed
Extracts ActivityItem into a shared component and reuses it on the
client detail page, filtered to the current client, so "History Book"
now shows activity instead of a free-text notes editor.
